+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sopt-mds/ui
 
+## 1.0.5
+
+### Patch Changes
+
+- 3caf2a1: register checkbox component and add fg.neutral.ghostDisabled token
+- Updated dependencies [3caf2a1]
+  - @sopt-mds/design-tokens@1.0.4
+
+## 1.0.4
+
+### Patch Changes
+
+- 0b287ce: register action button component
+- 0b287ce: register floating button component
+- 0b287ce: register reaction button component
+- 0b287ce: register text button component
+
 ## 1.0.3
 
 ### Patch Changes